Refactor SdLoraTab into a general PipelineTab.
